flask_messages_gen/app/requirements.txt
kafka-python==1.4.6
Flask==1.0.3
